requirements.txt
streamlit
langchain
langchain-community
langchainhub
langgraph
langchain-openai
python-dotenv
sqlalchemy
pandas
matplotlib
numpy
pymongo
openai